MLFlow_Labs/Lab2/requirements.txt
# Core Dependencies
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq

# MLflow and Tracking
mlflow

# Vector Store and Embeddings
sentence-transformers==3.3.1
chromadb==0.5.23
faiss-cpu==1.9.0

# Document Processing
pypdf==5.1.0
pymupdf==1.24.14
arxiv==2.1.3
python-dotenv==1.0.1

# Utilities
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1

# Evaluation
rouge-score==0.1.2
bert-score==0.3.13

# Jupyter (for notebook)
jupyter==1.1.1
ipykernel==6.29.5